Moving SQL and PPL Provider Registration to Query Enhancements Plugin
Issue Resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/9584
