+++ src/plugins/query_enhancements/public/autocomplete/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { getSQLSuggestions } from './sql_suggestions';
+export { getPPLSuggestions } from './ppl_suggestions';
